Fixed missed mutation of instructions that caused new invalid instructions to be added to a robots program.  Now getting a new random instruction from the mutator returns a pre-mutated instruction.
Added instruction prevalence support.  Instructions can return a weight
from 0.0 to 1.0 to determine how likely they are to be used.  With this
we can tune the ratio of instructions to each other.  Initial example
modifies avgofcolumn,percentileofcolumn,highestofcolumn (etc) to return
1.0 the first time the instruction is added to a robot and a declining
weight for each additional attempt to add the same instruction to a
robot.  After an instruction is chosen, a call to nextDouble must score
less than the prevalance for the instruction to be allowed.

Robots re-factored to correctly call base class constructors and
mutators.  Removes need to instruction constructor to be aware of super
class data elements and relies on inheritance to correctly copy objects.